[Refactor] pages폴더에서 components폴더로 컴포넌트 이동
