+++ src/components/Community/PostDetail.tsx
