+++ src/components/Community/SearchStocktype.tsx
